Modifies CLI to show nodes by instance uuid
This patch updates the node-show subcommand for ironic so that it
shows the node that matches the given instance UUID.
ironic node-show [--instance] id
It also updates node-list subcommand so that it can show all the
nodes that have instance UUID associated or the opposite.
ironic node-list [--associated true|false]
This patch also fixes a few strings for i18n support.

Change-Id: I2aec9ba6e296e0d1db9f3ef9eed7236ecfb32cf6
Closes-Bug: #1239026