Shows 'last_error' property for a node
https://review.openstack.org/#/c/54466/ introduced a 'last_error'
field for a node. This string field is used to store the last error
from any requested asynchronous operation (eg, whether that was to
change power state, or deploy a node, or anything else) that
started but failed to finish. It is cleared when any new asynchronous
operation is started.

With this change, an 'ironic node-show' command will also return the
'last_error' value (if any) for the specified node.

Change-Id: I9663662a1a1237cc356e17947aa86b0bf0954339