Node power state is not printed after set
As of commit 4676d823a8a35 to Ironic, the API for changing Node power
state no longer returns a body. With no return body, there is nothing
to print in the CLI after requesting the power state change.

This patch removes the print-out of an empty table when the CLI
set-node-power-state command is issued.

Change-Id: If1fb035bdf83e36e3a6a13e639efadc370c40887
Related-bug: #1270160