Add set_provision_state command
The function of provisioning of the node is implemented as Web API and library,
but not exposed on the CLI.
This patch add set_provision_state command for 'active' and 'deleted' states.
The detail of set_provision_state command is below:

usage: ironic node-set-provision-state <node id> <provision state>
Provision the node active or deleted.
Positional arguments:
  <node id>          UUID of node
  <provision state>  Supported states: 'active' or 'deleted'

Change-Id: I2482b424e5d2806096b79cf81037db8cdd51bd63
Partial-Bug: #1310446