Fix misspelled class name AmbigiousAuthSystem
Class name AmbigiousAuthSystem in exc.py is misspelled.
To fix the issue, this patch renames AmbigiousAuthSystem to
AmbiguousAuthSystem in exc.py and creates an alias for compatibility.
Also edits the use of the class in client.py accordingly and adds an
additional assertRaises() to
test_client.py.test_get_client_no_url_and_no_token to test both the
correct exception and the alias.

Change-Id: I18309b045d5eb9e81c72b1ed9de350925927db66
Closes-Bug: 1356512