Correct node CREATION_ATTRIBUTE "uuid"
Ironic's REST API allows a client to supply a UUID when registering a node.
However, the client was incorrectly passing this value as the "node_id"
property, rather than the "uuid" property, resulting in the API service
ignoring the parameter and generating a new UUID anyway.

This corrects the behavior by passing the appropriately named parameter
if the user specified a UUID when registering a new node.

Closes-bug: #1381203
Change-Id: I23ebbf20d836f0afc507be5c9dbbe1d759a3cddb