Fix the usage comment of node-set-power-state cmd
Current usage comment of node-set-power-state cmd is below:

$ ironic help node-set-power-state
usage: ironic node-set-power-state <node id> <power state>

Power the node on or off.

Positional arguments:
  <node id>      UUID of node
  <power state>  Supported states: 'on' or 'off' or 'reboot'

usage comment "Power the node on or off." and
arguments description
"<power state>  Supported states: 'on' or 'off' or 'reboot'"
are different, so this lacks consistency.
This patch set fix this usage comment to
"Power the node on or off or reboot."

Change-Id: Idd5320a890f5fdd3b68a83573c5898da293c4583