Fix node-set-provision-state cmd's help strings
Current help strings of node-set-provision-state cmd is below:

$ ironic help node-set-provision-state
usage: ironic node-set-provision-state <node id> <provision state>

Provision an instance on, or delete an instance from a node.

Positional arguments:
  <node id>          UUID of node
  <provision state>  Supported states: 'active' or 'deleted' or 'rebuild'

Help strings "Provision an instance on, or delete an instance from a node."
and arguments description
"<provision state>  Supported states: 'active' or 'deleted' or 'rebuild'"
are different because we can also do rebuild.
This patch set fix this help strings to
"Provision, rebuild or delete an instance."

Change-Id: I21555a9dfc46b5539749de47e0cd4d403d8cd302