Capture stdout or stderr in tests
This uses the environment variables OS_STDOUT_CAPTURE and
OS_STDERR_CAPTURE to determine whether or not to capture
stdout and stderr (respectively) in the unit tests. These
will appear in the test results for tests that fail.

Capturing will be done if the value is set to one of
(case-insensitive) '1', 't', 'true', 'on', 'y', 'yes'.
Any other value will turn off capturing. The default (via
.testr.conf) is to capture.

Change-Id: I3e6006aa1ed1cf3fad7e55a7201be1c10d165431
Closes-Bug: #1421307