Fix pep8 stderr warning regarding __all__ defined as list
During the pep8 run get the following warning generated:

./ironicclient/__init__.py WARNING: __all__ is defined as a list, this
means pep257 cannot reliably detect contents of the __all__ variable,
because it can be mutated. Change __all__ to be an (immutable) tuple, to
remove this warning. Note, pep257 uses __all__ to detect which
definitions are public, to warn if public definitions are missing
docstrings. If __all__ is a (mutable) list, pep257 cannot reliably
assume its contents. pep257 will proceed assuming __all__ is not
mutated.

Change-Id: I25a5d24e9c429bdd119abf9efb24e3cc822eef4a