Add invalid attribute names to exception
When creating a resource via CreateManager.create(),
an InvalidAttribute could be raised for invalid attributes
(attributes that are not needed to create the resource).
The raised exception had no message.

This changes the exception to include a list of all the
invalid attributes as well as indicate what the problem
is.

Change-Id: Iaea056ea6946a6f813f8081c26f62dcfe63d1bbd