Fix display of chassis UUID field if empty
When calling OSC commands "openstack baremetal node create" or
"openstack baremetal node show" the field "chassis_uuid" should be
displayed even if it is empty.

This patch also updated the test schema for the ironic --json node-*
command as chassis_uuid will now appear as 'null' when the node isn't
associated with a chassis.

Change-Id: Ic10eb9ef908c12d5c71c1df407bfb66079a25703
Closes-Bug: #1620749