Bump the LAST_KNOWN_API_VERSION to 29 in OSC
The command to inject NMI has been merged for OSC [0] but the API version
required for it hasn't been bumped. This patch fixes the problem.

Note that this patch has not been cherry-picked, that's because this
problem has been fixed by another patch [1] in the master branch which
is bumping the maximum version to 31.

[0] https://review.openstack.org/#/c/357661/
[1] https://review.openstack.org/#/c/418359/

Closes-Bug: #1662959
Change-Id: I285178d0b7384956eb151ca66007d7354566574d