Do not depend on python-openstackclient
OpenStack client is not a runtime dependency of ironicclient,
and having it in requirements just brings in many dependencies
which might not be needed at all when using the Python API of the client
only (for example in server-side applications).

Although dependency on osc-lib is enough for unit tests, add
python-openstackclient to test-requirements so that functional tests
pass.

Also, add a setuptools 'extra' so that users can install
python-openstackclient together with ironicclient if wishing to do so
as follows:

    pip install python-ironicclient[cli]

Change-Id: Ic7d06e61cd234b327613287802361c58bf6bf11e
Closes-Bug: #1562023