Log warning when API version is not specified for the OSC plugin
At the Pike PTG, an issue was brought up regarding the use of an old API
version in the ironic OpenStack CLI plugin. [0]

It was suggested that we begin printing a warning whenever the default
OSC API version is used and later default to using the latest API
version when --os-baremetal-api-version is unspecified.

This patch adds the warning discussed above.

[0] https://etherpad.openstack.org/p/ironic-pike-ptg-operations L30

Co-Authored-By: Dmitry Tantsur <dtantsur@redhat.com>
Partial-Bug: #1671145
Change-Id: I0cf4e737595405b7f9beff76a72ffd26e07e6277