Allow OS_BAREMETAL_API_VERSION=latest to work
Allow OS_BAREMETAL_API_VERSION=latest to work. If
OS_BAREMETAL_API_VERSION=latest then convert to the latest version and
set that as the 'default' value for the os_baremetal_api_version
value.

Remove the LOG.debug() message about setting to latest as the message
never appears in the logfile. As tested with:
  $ openstack --debug --log-file log.txt baremetal node list --os-baremetal-api-version latest

Also it was set to always log the message.

Change-Id: Idc3de8ae55e8267342633668a06d18b45c9e5f0a
Closes-Bug: #1712935