Allow to use none auth in functional tests
This change will make possible adding a new functional job running
ironic in noauth mode, and accessing it with none auth plugin from
the tests. There was no decision if such job is needed, but still
seems to be a good thing to keep the code updated to correspond to
what is the intended way of working with noauth ironic.

Change-Id: I48cf37c87fdb74a3b38742a929698c9bd146d7d1