Do not use urljoin in base http client
this fails when ironic API endpoint is not in the form
"host:port", but "host/vhost" instead
(as when ironic-api is deployed behind Apache),
since the 'vhost' gets swallowed by 'urljoin'.

This leads to a failure when using ironicclient with token and
endpoint passed in, otherwise a SessionClient is used that does
not have this problem.

Simply concat those url parts together (ensuring there is at least a
single '/' between them).

Change-Id: I583e0f9bdc81a655861c6ff508782173021428f0
Closes-Bug: #1721599