Move legacy ironicclient jobs in-tree
This moves the legacy python-ironicclient CI jobs into the
python-ironicclient tree, instead of storing them in
openstack-infra/openstack-zuul-jobs.

This also changes the ironicclient-tempest-dsvm-src job to
inherit from legacy-ironic-dsvm-base instead of legacy-dsvm-base,
since the required projects are the same.

This will give us control as we migrate the legacy jobs to the
new ansible roles.

Change-Id: If0c072e357fbeccc3dee8cc06b96e1bfa42299a7