Facilitate latest Rest API use
In order to provide insight into the remote
API verison, we need the ability to negotiate
upon the latest API version available, and then
report what that version is.

In order to understand if this has occured,
we also need to provide insight into if version
negotiation has occured.

Adds logic to the session/http clients to faciltate
version negotiation on the latest available version,
and provide user insight into that verison.

Change-Id: I813237eee4b122211f95558f677b25e0675569d5
Related-Bug: #1739440
Related-Bug: #1671145