fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Change-Id: I77dfc7cc757befdc988016f5f90e4dddada5cc69
Signed-off-by: Doug Hellmann <doug@doughellmann.com>