pass endpoint interface to http client
The 'interface' argument was being ignored so that the HTTP client was
always using the public endpoint for Ironic. This fixes it so that the
'interface' argument is taken into consideration.

There's also no need to explicitly set the interface to 'publicURL'
because that's already the default in keystoneauth.

Change-Id: I610836e5038774621690aca88b2aee25670f0262
story: 2005118
task: 29802
(cherry picked from commit d6eea403cbf0d11b06acbecc704f422a7e278462)