Use endpoint_override in version negotiation
Change [1] introduced the usage of endpoint_filter when deciding which
endpoint to use during version negotiation. This requires properly
setting interface/region/service_type in order to properly pick the
endpoint. But, sometimes (for example in Kolla-ansible), the endpoint
url is explicitly defined. In the event of the endpoint being explicitly
defined, it should take priority here.

It should be noted that even though SessionClient extends from
keystoneauth1's adapter.LegacyJsonAdapter, the `request` function within
_make_simple_request is being called on the nested Session object, not
the Adapter itself. If it was called on the Adapter this problem
wouldn't exist, as the `endpoint_override` (or `endpoint`) kwarg passed
in to the constructor of the SessionClient would already take
precedence.

Additionally: Adds a release note to the older branch as this is a
fix that needs to be released and is worthy of change visibility
in release notes.

[1]: I42b66daea1f4397273a3f4eb1638abafb3bb28ce

Change-Id: I69dc2c88648ba1d09a9ad3ab3435662e8d1ea6ff
Related-Bug: #1818295
(cherry picked from commit 8071d5de57c728ce09d25d04c8c9fe48cd7db42e)
(cherry picked from commit b40630f3b3e747ae4a26d31e8e292554b99713f2)