Get rid of the oslo.serialization dependency
It was used to smoothen out the difference in handling bytes in
Python 2 and 3. Now that we only support Python 3, it can be
replaced.

A side effect of this change is that we no longer accept bytes
in JSON. JSON does not support bytes, but this problem has been
hidden due to oslo.serialization until now. The configdrive
handling has been updated to account for that.

Change-Id: I230b55db55bce08d46f5023ad7a3f6501c96d100