Remove use of distutils
This has been removed in Python 3.12, which is preventing us using
ironicclient in those environments. Thankfully we only use it for
StrictVersion which we are using to parse microversions. This is
actually overkill here since we can always expect a version string like
'1.5', thus we do this parsing ourselves rather than drag in a new
library like 'microversion_parse'.

Change-Id: I7630ecfa2797c502868994302926bd2ac07d1ff7
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>