bump minimum pbr version for pip 23.1 support
While pbr 5.7.0 is the minimum version required, due to bug fixes we've
settled on 6.0.0 for all other ironic components so bump to that. Bump
the requirements.txt and setup.py to match since the versions mentioned
there are too old. Remove wheel as it's not actually required for the
build system.

Change-Id: I9359d967bdc5ac2a66b5bf826de5151e33a89b1d
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>