Fix parallel initial version negotiation
If two parallel greenthreads use the same, uninitialized client, there's
a race-condition when both enter `negotiate_version()`: there is a
request made to Ironic, which hands over execution to the other
greenthread. The first greenthread returning from the request changes
the state of the client and the second one reads the updated state and
thus thinks it's in an error-handling call instead of the initial
negotiation - and errors out.

We fix this by adding a lock around the initial call to
`negotiate_version()`.

Change-Id: I9ec03d2bc34017c7670fd6903e5353a8c91e9f17
Closes-Bug: #2119323
Signed-off-by: Johannes Kulik <johannes.kulik@sap.com>
(cherry picked from commit ad1291cc9cb4fc6560a1f82d73a496168a6770c6)